Finally, address the hidden call graph
If one of the C files calls a function we don't have the source of,
(e.g. ep_FOO)  and this function in turn calls ep_BAR, then ep_BAR
is NOT a dead function - even though it does not appear anywhere
in any of the C sources we process.

So we will also collect the mentions made in the disassembly
of the ELF.