Avoid reporting "phantom" symbols.
One would expect the list of Function symbols reported by objdump -t and
filtering for "F .text" to be the perfect source of truth in terms of
the complete set of functions in a binary.

Let's call this list the A list.

Very surprisingly, the A list contains symbols that exist *nowhere*
in the disassembled output.

I really mean nowhere - no definition, no reference, nothing.

Looking up their reported addresses - from the objdump output - we instead
find some other function in there.

In the final stage of the logic, when we scout the disassembled output
for references to symbols, since we don't find any reference to these
"phantoms", we were so far reporting them as "dead code".

So to clean up the list of dead functions from these "phantoms",
we now collect the symbols that appear as "<symbolName>:" in the
disassembled output, and intersect it with the A list. What remains
is the "true" A list - one that doesn't have these mysterious symbols.

Preliminary research seems to indicate these are linker-aliased symbols;
why we use them, remains for now a mystery.